+++ components/overview-body/overview-body.acss
@@ -1,0 +1,84 @@
+.item-header {
+  height: 32rpx;
+  padding: 8rpx;
+  display: flex;
+  justify-content: space-around;
+}
+
+.swiper-container-item {
+  display: flex;
+  position: relative;
+  flex-direction: column;
+  margin: 5rpx 20rpx;
+  padding: 5rpx;
+  border: gray solid 5rpx;
+  border-radius: 15rpx;
+  height: 240rpx;
+  overflow: hidden;
+  transition: all 0.5s;
+}
+
+.item-destination-arrow::before {
+  content: "";
+  display: inline-block;
+  width: 40rpx;
+  height: 40rpx;
+  margin-left: 5rpx;
+  background-image: url("/images/arrow.svg");
+  background-size: contain;
+}
+
+.item-bus-time {
+  font-size: 28rpx;
+  line-height: 28rpx;
+  vertical-align: middle;
+}
+
+.item-destination-name {
+  margin-left: 10rpx;
+  margin-right: 10rpx;
+  padding: 5rpx;
+  border: gray solid 3rpx;
+  border-radius: 10rpx;
+  text-align: center;
+  font-size: 26rpx;
+}
+
+.item-bus-line {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  vertical-align: middle;
+}
+
+.item-bus-line-detail {
+  height: 100%;
+  width: 100%;
+  overflow-y: auto;
+  margin-top: 20rpx;
+  padding: 0 20rpx;
+}
+
+.item-bus-remark {
+  display: block;
+  position: absolute;
+  background-color: #fff;
+  z-index: 20;
+  bottom: 5rpx;
+  width: 100%;
+  height: 30rpx;
+  font-size: 24rpx;
+  margin: 0 5rpx;
+  white-space: nowrap;
+  overflow: auto;
+}
+
+.item-bus-name {
+  max-width: 200rpx;
+  font-size: 28rpx;
+  line-height: 28rpx;
+  white-space: nowrap;
+  overflow-y: hidden;
+  overflow-x: auto;
+  vertical-align: middle;
+}
