+++ components/overview-body/overview-body.js
@@ -1,9 +1,13 @@
+import { bodyData, bodyProps } from "/options/props/overview";
+import { methods } from "./handlers";
+import observers from "./observers";
+
 Component({
-  mixins: [],
-  data: {},
-  props: {},
-  didMount() {},
-  didUpdate() {},
-  didUnmount() {},
-  methods: {},
+  options: {
+    observers: true,
+  },
+  data: bodyData,
+  props: bodyProps,
+  observers,
+  methods,
 });
