+++ components/overview-body/overview-body.json
@@ -1,4 +1,7 @@
 {
   "component": true,
-  "styleIsolation": "apply-shared"
+  "styleIsolation": "apply-shared",
+  "usingComponents": {
+    "steps": "/components/steps/steps"
+  }
 }
