+++ components/overview-header/overview-header.acss
@@ -1,161 +1,108 @@
 .item-header {
-    height: 32rpx;
-    padding: 8rpx;
-    display: flex;
-    justify-content: space-around;
-  }
-  
-/* -----overview-------- */
+  height: 32rpx;
+  padding: 8rpx;
+  display: flex;
+  justify-content: space-around;
+}
 
 .item-destinations {
-    font-size: 28rpx;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .item-destination-name {
-    margin-left: 10rpx;
-    margin-right: 10rpx;
-    padding: 5rpx;
-    border: gray solid 3rpx;
-    border-radius: 10rpx;
-    text-align: center;
-    font-size: 26rpx;
-  }
-  
-  .item-destination-title-name {
-    height: 32rpx;
-    font-size: 30rpx;
-    line-height: 32rpx;
-    font-weight: bold;
-    margin-left: 10rpx;
-    margin-right: 10rpx;
-    padding: 5rpx;
-    border: gray solid 3rpx;
-    border-radius: 10rpx;
-    text-align: center;
-  }
-  
-  .item-destination-cancel {
-    height: 28rpx;
-    font-size: 26rpx;
-    line-height: 28rpx;
-    margin-left: 10rpx;
-    margin-right: 10rpx;
-    padding: 5rpx;
-    border: gray dashed 3rpx;
-    border-radius: 10rpx;
-    text-align: center;
-  }
-  
-  .item-destination-cancel:before {
-    content: "\2716";
-    color: gray;
-  }
-  
-  .item-destination-name.end-active {
-    background-color: #e5e5e5;
-  }
-  
-  .item-bus-time {
-    font-size: 28rpx;
-    line-height: 28rpx;
-    vertical-align: middle;
-  }
-  
-  .item-destination-arrow::before {
-    content: "";
-    display: inline-block;
-    width: 40rpx;
-    height: 40rpx;
-    margin-left: 5rpx;
-    background-image: url("/images/arrow.svg");
-    background-size: contain;
-  }
-  
-  .item-switch-direction {
-    display: flex;
-    align-items: center;
-  }
-  
-  .item-switch-direction::before {
-    content: "";
-    display: inline-block;
-    width: 40rpx;
-    height: 40rpx;
-    margin-right: 10rpx;
-    background-image: url("/images/switch.svg");
-    background-size: contain;
-  }
-  
-  .item-body-main {
-    position: relative;
-    height: 85%;
-    margin: 15rpx;
-    border-bottom: solid gray 5rpx;
-    box-sizing: content-box;
-  }
-  
-  .item-bus-object {
-    position: absolute;
-    bottom: 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .item-bus-caption {
-    position: absolute;
-    bottom: 80rpx;
-    font-size: 26rpx;
-    white-space: nowrap;
-  }
-  
-  .item-stop-caption {
-    position: absolute;
-    bottom: 110rpx;
-    font-size: 26rpx;
-    white-space: nowrap;
-  }
-  
-  .item-bus-icon {
-    content: "";
-    display: inline-block;
-    background-size: contain;
-  }
-  
-  .item-bus-icon.bus {
-    background-image: url("/images/bus.svg");
-    width: 120rpx;
-    height: 120rpx;
-    translate: 0 40rpx;
-  }
-  
-  .item-bus-icon.babybus {
-    background-image: url("/images/babybus.svg");
-    width: 100rpx;
-    height: 100rpx;
-    translate: 0 25rpx;
-  }
-  
-  .item-bus-icon.shuttle {
-    background-image: url("/images/shuttle.svg");
-    width: 90rpx;
-    height: 90rpx;
-    translate: 0 20rpx;
-    scale: -1 1;
-  }
-  
-  .item-bus-icon-flag {
-    background-image: url("/images/flag.svg");
-    width: 60rpx;
-    height: 60rpx;
-  }
-  
-  .item-bus-icon-human-walk {
-    background-image: url("/images/walk.svg");
-    width: 60rpx;
-    height: 60rpx;
-  }
+  font-size: 28rpx;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.item-destination-title-name {
+  height: 32rpx;
+  font-size: 30rpx;
+  line-height: 32rpx;
+  font-weight: bold;
+  margin-left: 10rpx;
+  margin-right: 10rpx;
+  padding: 5rpx;
+  border: gray solid 3rpx;
+  border-radius: 10rpx;
+  text-align: center;
+}
+
+.item-destination-cancel {
+  height: 28rpx;
+  font-size: 26rpx;
+  line-height: 28rpx;
+  margin-left: 10rpx;
+  margin-right: 10rpx;
+  padding: 5rpx;
+  border: gray dashed 3rpx;
+  border-radius: 10rpx;
+  text-align: center;
+}
+
+.item-destination-cancel:before {
+  content: "\2716";
+  color: gray;
+}
+
+.item-body-main {
+  position: relative;
+  height: 85%;
+  width: 100%;
+  margin: 15rpx;
+  border-bottom: solid gray 5rpx;
+  box-sizing: content-box;
+}
+
+.item-bus-object {
+  position: absolute;
+  bottom: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.item-bus-caption {
+  position: absolute;
+  bottom: 80rpx;
+  font-size: 26rpx;
+  white-space: nowrap;
+}
+
+.item-bus-icon {
+  content: "";
+  display: inline-block;
+  background-size: contain;
+}
+
+.item-bus-icon.bus {
+  background-image: url("/images/bus.svg");
+  width: 120rpx;
+  height: 120rpx;
+  translate: 0 40rpx;
+}
+
+.item-bus-icon.babybus {
+  background-image: url("/images/babybus.svg");
+  width: 100rpx;
+  height: 100rpx;
+  translate: 0 25rpx;
+}
+
+.item-bus-icon.shuttle {
+  background-image: url("/images/shuttle.svg");
+  width: 90rpx;
+  height: 90rpx;
+  translate: 0 20rpx;
+  scale: -1 1;
+}
+
+.item-bus-icon-flag {
+  background-image: url("/images/flag.svg");
+  width: 60rpx;
+  height: 60rpx;
+}
+
+.item-bus-icon-human-walk {
+  background-image: url("/images/walk.svg");
+  width: 60rpx;
+  height: 60rpx;
+}
