+++ options/apis/carApis.js
@@ -24,6 +24,7 @@ const funcMap = {
 
 export async function queryBackend(dataName, dataType, parm, show=true) {
   if (show) showQuerying(dataType);
+  console.log(dataName, dataType, parm);
   const ret = await funcMap[dataName][dataType](...parm);
   if (show) hideLoading();
   return ret;
